defguard::db::models::device

Struct Device

pub struct Device<I = NoId> {
    pub id: I,
    pub name: String,
    pub wireguard_pubkey: String,
    pub user_id: Id,
    pub created: NaiveDateTime,
    pub device_type: DeviceType,
    pub description: Option<String>,
    pub configured: bool,
}

Fields
id: Iname: Stringwireguard_pubkey: Stringuser_id: Idcreated: NaiveDateTimedevice_type: DeviceTypedescription: Option<String>configured: bool

Whether the device should be considered as setup and ready to use or does it require some additional steps to be taken. Not configured devices won’t be sent to the gateway. It is assumed that an unconfigured device is already added to all networks it should be in, but it’s not ready to be used yet due to e.g. public key not properly set up yet.

Implementations

impl Device<NoId>

pub async fn save<'e, E>(self, executor: E) -> Result<Device<Id>, Error>
where E: PgExecutor<'e>,

pub fn with_id(self, id: Id) -> Device<Id>

impl Device<Id>

pub async fn find_by_id<'e, E>( executor: E, id: Id, ) -> Result<Option<Self>, Error>
where E: PgExecutor<'e>,

pub async fn all<'e, E>(executor: E) -> Result<Vec<Self>, Error>
where E: PgExecutor<'e>,

pub async fn delete<'e, E>(self, executor: E) -> Result<(), Error>
where E: PgExecutor<'e>,

pub async fn save<'e, E>(&mut self, executor: E) -> Result<(), Error>
where E: PgExecutor<'e>,

pub fn as_noid(self) -> Device

impl Device

pub fn new( name: String, wireguard_pubkey: String, user_id: Id, device_type: DeviceType, description: Option<String>, configured: bool, ) -> Self

impl Device<Id>

pub async fn add_to_all_networks( &self, transaction: &mut PgConnection, ) -> Result<(Vec<DeviceNetworkInfo>, Vec<DeviceConfig>), DeviceError>

pub async fn find_network_device_networks<'e, E>( &self, executor: E, ) -> Result<Vec<WireguardNetwork<Id>>, SqlxError>
where E: PgExecutor<'e>,

Gets the first network of the network device FIXME: Return only one network, not a Vec

pub fn validate_pubkey(pubkey: &str) -> Result<(), String>