Struct defguard::db::models::user::User

pub struct User {
    pub id: Option<i64>,
    pub username: String,
    pub last_name: String,
    pub first_name: String,
    pub email: String,
    pub phone: Option<String>,
    pub mfa_enabled: bool,
    pub is_active: bool,
    /* private fields */
}

Fields
id: Option<i64>username: Stringlast_name: Stringfirst_name: Stringemail: Stringphone: Option<String>mfa_enabled: boolis_active: bool
Implementations

impl User

pub async fn start_enrollment( &self, transaction: &mut PgConnection, admin: &User, email: Option<String>, token_timeout_seconds: u64, enrollment_service_url: Url, send_user_notification: bool, mail_tx: UnboundedSender<Mail> ) -> Result<String, TokenError>

Start user enrollment process This creates a new enrollment token valid for 24h and optionally sends enrollment email notification to user

pub async fn start_remote_desktop_configuration( &self, transaction: &mut PgConnection, admin: &User, email: Option<String>, token_timeout_seconds: u64, enrollment_service_url: Url, send_user_notification: bool, mail_tx: UnboundedSender<Mail> ) -> Result<String, TokenError>

Start user remote desktop configuration process This creates a new enrollment token valid for 24h and optionally sends email notification to user

impl User

pub async fn find_by_id<'e, E>( executor: E, id: i64 ) -> Result<Option<Self>, Error>
where E: PgExecutor<'e>,

pub async fn all<'e, E>(executor: E) -> Result<Vec<Self>, Error>
where E: PgExecutor<'e>,

pub async fn delete<'e, E>(self, executor: E) -> Result<(), Error>
where E: PgExecutor<'e>,

pub async fn save<'e, E>(&mut self, executor: E) -> Result<(), Error>
where E: PgExecutor<'e>,

impl User

pub fn new<S: Into<String>>( username: S, password: Option<&str>, last_name: S, first_name: S, email: S, phone: Option<String> ) -> Self

pub fn set_password(&mut self, password: &str)

pub fn verify_password(&self, password: &str) -> Result<(), HashError>

pub fn has_password(&self) -> bool

pub fn name(&self) -> String

pub async fn new_totp_secret<'e, E>( &mut self, executor: E ) -> Result<String, SqlxError>
where E: PgExecutor<'e>,

Generate new TOTP secret, save it, then return it as RFC 4648 base32-encoded string.

pub async fn new_email_secret<'e, E>( &mut self, executor: E ) -> Result<(), SqlxError>
where E: PgExecutor<'e>,

Generate new email secret, similar to TOTP secret above, but don’t return generated value.

pub async fn set_mfa_method<'e, E>( &mut self, executor: E, mfa_method: MFAMethod ) -> Result<(), SqlxError>
where E: PgExecutor<'e>,

pub async fn verify_mfa_state(&mut self, pool: &DbPool) -> Result<(), WebError>

Verify the state of mfa flags are correct. Recovers from invalid mfa_method Use this function after removing any of the authentication factors.

pub async fn enable_mfa(&mut self, pool: &DbPool) -> Result<(), WebError>

Enable MFA. At least one of the authenticator factors must be configured.

pub async fn get_recovery_codes<'e, E>( &mut self, executor: E ) -> Result<Option<Vec<String>>, SqlxError>
where E: PgExecutor<'e>,

Get recovery codes. If recovery codes exist, this function returns None. That way recovery codes are returned only once - when MFA is turned on.

pub async fn disable_mfa(&mut self, pool: &DbPool) -> Result<(), SqlxError>

Disable MFA; discard recovery codes, TOTP secret, and security keys.

pub async fn enable_totp<'e, E>(&mut self, executor: E) -> Result<(), SqlxError>
where E: PgExecutor<'e>,

Enable TOTP

pub async fn disable_totp(&mut self, pool: &DbPool) -> Result<(), SqlxError>

Disable TOTP; discard the secret.

pub async fn enable_email_mfa<'e, E>( &mut self, executor: E ) -> Result<(), SqlxError>
where E: PgExecutor<'e>,

Enable email MFA

pub async fn disable_email_mfa( &mut self, pool: &DbPool ) -> Result<(), SqlxError>

Disable email MFA; discard the secret.

pub async fn all_without_sensitive_data( pool: &DbPool ) -> Result<Vec<UserDiagnostic>, SqlxError>

Select all users without sensitive data.

pub async fn find_by_group_name( pool: &DbPool, group_name: &str ) -> Result<Vec<User>, SqlxError>

Return all members of group

pub fn verify_totp_code(&self, code: u32) -> bool

Check if TOTP code is valid.

pub fn generate_email_mfa_code(&self) -> Result<u32, WebError>

pub fn verify_email_mfa_code(&self, code: u32) -> bool

Check if email MFA code is valid.

pub async fn verify_recovery_code( &mut self, pool: &DbPool, code: &str ) -> Result<bool, SqlxError>

Verify recovery code. If it is valid, consume it, so it can’t be used again.

pub async fn find_by_username<'e, E>( executor: E, username: &str ) -> Result<Option<Self>, SqlxError>
where E: PgExecutor<'e>,

pub async fn find_by_email<'e, E>( executor: E, email: &str ) -> Result<Option<Self>, SqlxError>
where E: PgExecutor<'e>,

pub async fn member_of_names<'e, E>( &self, executor: E ) -> Result<Vec<String>, SqlxError>
where E: PgExecutor<'e>,

pub async fn member_of<'e, E>( &self, executor: E ) -> Result<Vec<Group>, SqlxError>
where E: PgExecutor<'e>,

pub async fn devices(&self, pool: &DbPool) -> Result<Vec<UserDevice>, SqlxError>

pub async fn wallets<'e, E>( &self, executor: E ) -> Result<Vec<WalletInfo>, SqlxError>
where E: PgExecutor<'e>,

pub async fn oauth2authorizedapps<'e, E>( &self, executor: E ) -> Result<Vec<OAuth2AuthorizedAppInfo>, SqlxError>
where E: PgExecutor<'e>,

pub async fn security_keys( &self, pool: &DbPool ) -> Result<Vec<SecurityKey>, SqlxError>

pub async fn add_to_group<'e, E>( &self, executor: E, group: &Group ) -> Result<(), SqlxError>
where E: PgExecutor<'e>,

pub async fn remove_from_group<'e, E>( &self, executor: E, group: &Group ) -> Result<(), SqlxError>
where E: PgExecutor<'e>,

pub async fn remove_oauth2_authorized_apps<'e, E>( &self, executor: E, app_client_ids: &[i64] ) -> Result<(), SqlxError>
where E: PgExecutor<'e>,

Remove authorized apps by their client id’s from user

pub async fn init_admin_user( pool: &DbPool, default_admin_pass: &str ) -> Result<(), Error>

Create admin user if one doesn’t exist yet

pub async fn logout_all_sessions<'e, E>( &self, executor: E ) -> Result<(), SqlxError>
where E: PgExecutor<'e>,

impl User

pub fn from_searchentry( entry: &SearchEntry, username: &str, password: &str ) -> Self

pub fn as_ldap_mod(&self, config: &LDAPConfig) -> Vec<Mod<&str>>

pub fn as_ldap_attrs<'a>( &'a self, ssha_password: &'a str, nt_password: &'a str ) -> Vec<(&'a str, HashSet<&'a str>)>