defguard::enterprise::db::models

Module acl

Structs

AclAlias

Database representation of an ACL alias. Aliases can be used to define the destination part of an ACL rule so that it’s easier to create new rules with common restrictions. In addition to the AclAlias we provide AclAliasInfo and ApiAclAlias that combine all related objects for easier downstream processing.

AclAliasDestinationRange

AclAliasInfo

Helper struct combining all DB objects related to given AclAlias. All related objects are stored in vectors.

AclRule

Database representation of an ACL rule. ACL rule has many related objects:

AclRuleAlias

AclRuleDestinationRange

AclRuleDevice

AclRuleGroup

AclRuleInfo

Helper struct combining all DB objects related to given AclRule. All related objects are stored in vectors.

AclRuleNetwork

AclRuleUser

ParsedDestination

PortRange

Representation of port range. Those are stored in the db as [PgRange<i32>]. Single ports are represented as single-element ranges, e.g. port 80 = Range(80, 81) since upper bound is excluded by convention.

Enums

AclError

AliasKind

ACL alias can be of one of the following types:

AliasState

ACL alias can be in one of the following states:

RuleState

ACL rule can be in one of the following states:

Functions

parse_destination

Perses a destination string into singular ip addresses or networks and address ranges. We should be able to parse a string like this one: 10.0.0.1/24, 10.1.1.10-10.1.1.20, 192.168.1.10, 10.1.1.1-10.10.1.1

parse_ports

Perses a ports string into singular ports and port ranges We should be able to parse a string like this one: 22, 23, 8000-9000, 80-90

Type Aliases

Protocol

https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/tree/include/uapi/linux/in.h