defguard::enterprise::ldap

Module sync

This module contains the logic for synchronizing users and groups between Defguard and LDAP.

The synchronization is performed in two variants: full and incremental.

Sync status

The sync status is stored in the database and can be either InSync or OutOfSync. The status is used to determine whether the full sync should be performed or not. The status is set to OutOfSync when some Defguard changes couldn’t be propagated to LDAP (e.g. LDAP outage). The status is set to InSync when the sync is completed successfully.

Full synchronization

The full synchronization takes all objects (users, groups and their memberships) from one source, compares it with the other one and computes appropriate changes to make the two sources roughly equal.

The full sync is performed only when the sync status is set to OutOfSync.

The changes are computed with regard to a specified authority, which determines which source is considered to be the more important one and which is expected to be edited more often. The authority can be either LDAP or Defguard.

The authority has been introduced to solve the problem of ambiguity when some object is not present in one of the sources. Such scenario may occur when a user is deleted from one of the sources OR when a user is added to one of the sources. In each case, a different action should be taken to make the two sources equal (deletion or addition). For example:

User is not present in LDAP but is present in Defguard

Did we just add the user to Defguard but couldn’t propagate that change or did we delete the user from LDAP?

If the authority is LDAP, we should delete the user from Defguard, as we assume that it was more probable that the change was made in LDAP.

If the authority is Defguard, we should add the user to LDAP, as we assume that it was more probable that the change was made in Defguard.

If the LDAP connection is never lost and no other issues arise, the full sync should be performed only once, when the LDAP sync is enabled. So this is a more of a damage control mechanism rather than something that should be invoked regularly.

Incremental synchronization

The incremental synchronization is a regular synchronization operation which comes in two varieties: synchronous and asynchronous.

Changes from Defguard are propagated to LDAP in real-time, synchronously, to keep LDAP up-to-date with Defguard instantly. This is done by calling appropriate LDAP operations after each change in Defguard. Changes other way around (from LDAP to Defguard) are pulled asynchronously at regular intervals (every 5 minutes by default). Implementation-wise it’s done by running a full sync with LDAP authority, as it has the same effect when we consider that LDAP has the most recent Defguard changes (due to synchronous change propagation).

This synchronization should work reliably most of the time, given that:

LDAP connection is stable

The LDAP change pull is performed relatively often

One object is not changed in both sources between two asynchronous syncs (may cause overwriting of changes), but this sounds like an unlikely scenario

Potential improvements and issues

Some optimizations could be made using the implementation-specific object modification/creation timestamps in LDAP. Currently everything is compared as is, without any regard to the time of the change. We could skip some operations on objects that haven’t changed since the last sync. There is however still an issue with objects that have been deleted, LDAP doesn’t store deleted objects by default, so we may still need to compare full object lists.

There is no real pagination and everything is loaded into the memory at once. This may be an issue at some point. 10k LDAP records wasn’t a problem in testing. We may have bigger issues with other parts of Defguard with that user count though.

Enums

Authority

SyncStatus

Functions

get_ldap_sync_interval

get_ldap_sync_status

is_ldap_desynced

set_ldap_sync_status