defguard

Function init_vpn_location

pub async fn init_vpn_location(
    pool: &PgPool,
    args: &InitVpnLocationArgs,
) -> Result<String, Error>

Create a new VPN location. Meant to be used to automate setting up a new defguard instance. If the network ID has been specified, it will be assumed that the user wants to update the existing network or create a new one with a predefined ID. This is mainly used for deployment purposes where the network ID must be known beforehand.

If there is no ID specified, the function will only create the network if no other network exists. In other words, multiple networks can be created, but only if the ID is predefined for each network.