Module defguard::handlers

Modules

openid_flow

wireguard

worker

Structs

AddUserData

ApiResponse

Auth

AuthCode

AuthResponse

Return type needed to know if user came from openid flow with optional url to redirect him later if yes

AuthTotp

GroupInfo

PasswordChange

PasswordChangeSelf

RecoveryCode

RecoveryCodes

StartEnrollmentRequest

Username

WalletAddress

WalletChallenge

WalletChange

WalletSignature

WebAuthnRegistration

Functions

device_for_admin_or_self

Try to fetch [`Device’] if the device.id is of the currently logged in user, or the logged in user is an admin.

user_for_admin_or_self

Try to fetch User if the username is of the currently logged in user, or the logged in user is an admin.

Type Aliases

ApiResult