defguard::handlers

Function user_for_admin_or_self

pub async fn user_for_admin_or_self(
    pool: &PgPool,
    session: &SessionInfo,
    username: &str,
) -> Result<User<Id>, WebError>

Try to fetch User if the username is of the currently logged in user, or the logged in user is an admin.