Crate defguard

Modules

appstate

assets

auth

config

db

enterprise

globals

grpc

handlers

headers

hex

mail

secret

support

templates

updates

utility_thread

wg_config

wireguard_peer_disconnect

This module implements a functionality of disconnecting inactive peers in MFA-protected locations. If a device does not disconnect explicitly and just becomes inactive it should be removed from gateway configuration and marked as “not allowed”, which enforces an authentication requirement to connect again.

wireguard_stats_purge

Macros

global_value

Specify a global value that can be accessed from anywhere in the application. Positional arguments:

hashset

role

Constants

VERSION

Statics

SERVER_CONFIG

Functions

build_webapp

init_dev_env

Automates test objects creation to easily setup development environment. Test network keys: Public: zGMeVGm9HV9I4wSKF9AXmYnnAIhDySyqLMuKpcfIaQo= Private: MAk3d5KuB167G88HM7nGYR6ksnPMAOguAg2s5EcPp1M= Test device keys: Public: gQYL5eMeFDj0R+lpC7oZyIl0/sNVmQDC6ckP7husZjc= Private: wGS1qdJfYbWJsOUuP1IDgaJYpR+VaKZPVZvdmLjsH2Y=

init_vpn_location

Create a new VPN location. Meant to be used to automate setting up a new defguard instance. If the network ID has been specified, it will be assumed that the user wants to update the existing network or create a new one with a predefined ID. This is mainly used for deployment purposes where the network ID must be known beforehand.

run_web_server

Runs core web server exposing REST API.