defguard
Module handlers

Modules

openid_flow

wireguard

worker

Structs

AddUserData

ApiResponse

Auth

AuthCode

AuthResponse

Return type needed for knowing if a user came from OpenID flow. If so, fill in the optional URL field to redirect him later.

AuthTotp

EditGroupInfo

Dedicated GroupInfo variant for group modification operations.

GroupInfo

PasswordChange

PasswordChangeSelf

RecoveryCode

RecoveryCodes

StartEnrollmentRequest

Username

WalletAddress

WalletChallenge

WalletChange

WalletSignature

WebAuthnRegistration

WebHookData

Functions

device_for_admin_or_self

Try to fetch [`Device’] if the device.id is of the currently logged in user, or the logged in user is an admin.

user_for_admin_or_self

Try to fetch User if the username is of the currently logged in user, or the logged in user is an admin.

Type Aliases

ApiResult